+++ azure-kusto-data/tests/test_kusto_client.py
@@ -12,7 +12,7 @@ from pandas.util.testing import assert_frame_equal
 
 from azure.kusto.data.request import KustoClient
 from azure.kusto.data.exceptions import KustoServiceError
-from azure.kusto.data._response import WellKnownDataSet
+from azure.kusto.data.response import WellKnownDataSet
 
 
 def mocked_aad_helper():
